requirements.txt
discord.py==2.2.3
ffmpeg==1.4
PyNaCl==1.5.0